[b] 557 - Mobile landing page shrinks in height
The `full-height-wrapper` height was calculated inside render(). The bug
happened because render() was called before the browser’s viewport
could finish changing its size, leading to incorrect height value.

The fix is to only calculate the viewport size once on initial mount, at
which point things that mess up the viewport height (such as on-screen
keyboard) are not visible yet.

We could have gone with using `height: 100vh;` instead, but
unfortunately it completely ignores the browser address bar’s height.
This causes the address bar to overlap the landing page’s content when
it is visible. `Ignore address bar’s height` is actually expected
behavior [1] so there’s nothing we can do about it.

[#145022557]

[1]: https://bugs.chromium.org/p/chromium/issues/detail?id=428132