[c] 399 - Remove unnecessary ansible handler and notifications
We met this error while deploying staging: 

```
ERROR! The requested handler 'reload nginx' was not found in either the
main handlers list nor in the listening handlers list
```

Since we only need to restart nginx once at the end of the nginx setup
task, the handler is not really necessary. Therefore, it should be
removed to keep things simple.

[#149790399] https://www.pivotaltracker.com/story/show/149790399